fix(rust): Fix unicode length & dtype issues with array of null in lambda (#39)
* * correctly return the number of characters when calling length() in lambda. in case of UTF8, spark expects actual characters, but spark returns bytes.
* fix an issue when the dtype() of a series is incorrect when we use Array(Null) when inside lambda.
* add support for array(null) in more types in transform
* merged with PR:https://github.com/flarioncode/polars/pull/38/files

+++ crates/polars-core/src/series/implementations/mod.rs
@@ -569,7 +569,7 @@ mod tests {
 
         let transformed = transformed.unwrap();
         eprintln!("{:?}", transformed.dtype());
-        let result = transformed.i32();
+        let result = transformed.i64();
 
         assert!(result.is_ok());
 
@@ -578,7 +578,7 @@ mod tests {
 
     #[test]
     fn test_transform_numeric_different_type() {
-        // here we test that lambda can return complitly different type
+        // here we test that lambda can return completely different type
         let series =
             ChunkedArray::<Int32Type>::from_vec("array".into(), vec![1, 2, 3]).into_series();
 
@@ -628,7 +628,7 @@ mod tests {
 
         let transformed = transformed.unwrap();
         eprintln!("{:?}", transformed.dtype());
-        let result = transformed.i32();
+        let result = transformed.i64();
 
         assert!(result.is_ok());
 
@@ -652,7 +652,7 @@ mod tests {
 
         let transformed = transformed.unwrap();
         eprintln!("{:?}", transformed.dtype());
-        let result = transformed.i32();
+        let result = transformed.i64();
 
         assert!(result.is_ok());
